backend/requirements.txt
# Flask and dependencies
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0
